pyproject.toml
[tool.poetry]
name = "fii-aset"
version = "0.1.0"
description = ""
authors = ["Stoica Tedy <stoicatedy@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
Django = "^5.1.2"
celery = "^5.4.0"
django-extensions = "^3.2.3"
yfinance = "^0.2.48"
prometheus-client = "^0.21.0"
django-prometheus = "^2.3.1"
gunicorn = "^23.0.0"
psycopg2 = "^2.9.10"
django-tailwind = {extras = ["reload"], version = "^3.8.0"}
django-compress-staticfiles = "^1.0.1b0"
django-allauth-ui = "^1.5.1"
django-widget-tweaks = "^1.5.0"
slippers = "^0.6.2"
django-allauth = {extras = ["mfa"], version = "^65.1.0"}
requests = "^2.32.3"
django-stubs = {extras = ["compatible-mypy"], version = "^5.1.1"}
sentry-sdk = {extras = ["celery"], version = "^2.18.0"}
django-debug-toolbar = "^4.4.6"
django-csp = "^3.8"
django-permissions-policy = "^4.23.0"
django-money = {extras = ["exchange"], version = "^3.5.3"}
django-simple-history = "^3.7.0"


[tool.poetry.group.dev.dependencies]
pydotplus = "^2.0.2"
pre-commit = "^4.0.1"
poetry-plugin-export = "^1.8.0"
coverage = "^7.6.4"
celery-types = "^0.22.0"
werkzeug = "^3.1.2"
faker = "^30.8.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.coverage.run]
branch = true
source = ["backend"]
command_line = "backend/manage.py test"
relative_files = true

[tool.mypy]
plugins = ["mypy_django_plugin.main"]

[tool.django-stubs]
django_settings_module = "backend.settings.testing"
strict_settings = true
